--- src/components/EditProfileImage/EditProfileImage.stories.tsx
@@ -1,8 +1,0 @@
-import EditProfileImage from "@components/EditProfileImage";
-
-export default {
-  title: "component/EditProfileImage",
-  component: EditProfileImage,
-};
-
-export const editProfileImage = () => <EditProfileImage />;
