+++ src/pages/api/EditProfileAPI.ts
@@ -2,6 +2,7 @@ import HttpClient from "src/services/HttpClient";
 
 const EditProfileAPI = {
   getProfile: async (session: any) => {
+    if (session === undefined) return null;
     try {
       const path = `mypage/profile/${session?.userId}`;
       const response: [] | undefined = await HttpClient.get(
@@ -11,11 +12,11 @@ const EditProfileAPI = {
       );
       return response;
     } catch (e) {
-      console.log(e);
       return null;
     }
   },
   patchProfile: async (data: FormData, session: any) => {
+    if (session === undefined) return null;
     try {
       const path = `mypage/profile/${session?.userId}`;
       const response: [] | undefined = await HttpClient.patch(path, data, {
@@ -24,7 +25,6 @@ const EditProfileAPI = {
       });
       return response;
     } catch (e) {
-      console.log(e);
       return null;
     }
   },
